refactor(witness): move PR test modules to sibling files
Split the `#[cfg(test)] mod tests` blocks added by this PR out of
`partial_witness.rs` and `partial_witness_actor.rs` so reviewers can
collapse the test file and focus on production code. Same tests, same
count, no behavior change.

- `core/primitives/src/stateless_validation/partial_witness_tests.rs`
  (new) — 9 tests extracted from `partial_witness.rs`. No visibility
  changes required: tests exercise only the public API.
- `chain/client/src/stateless_validation/partial_witness/partial_witness_actor_tests.rs`
  (new) — 5 tests extracted from `partial_witness_actor.rs`. Marks
  `PENDING_V2_WITNESS_CACHE_SIZE`, `DeferOrigin`, `PendingV2WitnessEntry`
  (+ fields), and `PendingV2WitnessCache` (+ methods) as `pub(super)` so
  the sibling tests module can see them. Visibility is scoped to the
  parent `partial_witness` module — the crate-level API is unchanged.
- Wire both via `#[cfg(test)] mod <name>;` in the respective parent
  `mod.rs`, matching the `network_protocol/tests.rs` and
  `trie/trie_tests.rs` patterns already used in the repo.

Left alone:
- `chain/network/src/network_protocol/tests.rs` — already a tests-only
  file, no prod code to collapse.
- `chain/network/src/rate_limits/messages_limits.rs` — has 8 tests
  inline with prod, splitting only the one this PR added would create
  inconsistency.
