+++ chain/client/src/stateless_validation/partial_witness/mod.rs
@@ -1,6 +1,8 @@
 mod encoding;
 mod partial_deploys_tracker;
 pub mod partial_witness_actor;
+#[cfg(test)]
+mod partial_witness_actor_tests;
 mod partial_witness_tracker;
 
 pub use encoding::witness_part_length;
