+++ core/primitives/src/stateless_validation/mod.rs
@@ -7,6 +7,8 @@ pub mod chunk_endorsement;
 pub mod chunk_endorsements_bitmap;
 pub mod contract_distribution;
 pub mod partial_witness;
+#[cfg(test)]
+mod partial_witness_tests;
 pub mod spice_chunk_endorsement;
 pub mod spice_state_witness;
 pub mod state_witness;
